Update XMaterial.java to fix Terracotta
TERRACOTTA is being mapped to STAINED_CLAY when it should be HARDENED_CLAY for spigot versions v1.12.2 and earlier. When TERRACOTTA is parsed with STAINED_CLAY, it's resulting in WHITE_TERRACOTTA which is incorrect.  If it's HARDENED_CLAY then it will result in TERRACOTTA.  This issue occurs in spigot 1.12 and earlier.  For 1.13+ it uses TERRACOTTA so it's not an issue.